+++ mangadex_downloader/cli/url.py
@@ -71,7 +71,8 @@ def download_manga(url, args, legacy=False):
         args.use_chapter_title,
         args.unsafe,
         args.no_verify,
-        args.range
+        args.range,
+        args.force_https
     )
 
     if legacy:
@@ -96,7 +97,8 @@ def download_chapter(url, args, legacy=False):
         args.no_group_name,
         args.use_chapter_title,
         args.unsafe,
-        args.no_verify
+        args.no_verify,
+        args.force_https
     )
 
     if legacy:
@@ -134,7 +136,8 @@ def download_list(url, args):
         args.group,
         args.use_chapter_title,
         args.unsafe,
-        args.no_verify
+        args.no_verify,
+        args.force_https
     )
 
 # Legacy support